Add hint when displaying pre-released package if an update exists
Before this commit when querying for an installed package using pip if
the package that was returned contained a pre-released version of the
package it was not intuitive that the pre-released version of the
package could not be directly upgraded to without the `--pre` flag.

This commit shows additional information on how to upgrade to
pre-released versions of a package if one is found in the search.

Fixes #5169.