Give specific error on connnection failure
Pip gives a generic error message at the moment, if it is unable to
connect to the internet. This same error message is given if pip tries
to download a package which does not exist. This commit gives a specific
error message if pip is unable to connect to the server, to make clear
what the error is.