Merge of patch accepted in upstream msgpack-python/master improve legacy and jython support pending release.
https://github.com/msgpack/msgpack-python/commit/70b5f21b34b28cd50125ee337b5a399001759b73"

Note: I've also identified a zlib decompression issue in jython that
effects pip on jython 2.7.1. I submitted a patch to jython upstream.
Yes, it is still actively developed and embedded in some fairly
large java based platforms.

It plays more with the kotlin / rhino / groovy / nashorn arena, and the
main programmer focuses on programming to the exclusion of promotion and
organizing a community around it.