Add --target-scripts to provide path for binaries
For `pip install`, this introduces a new option, `--scripts-target`,
only to be used with `--target`, so that users can now specify a
different path for scripts/binaries instead of the default `$target_dir/bin`.

For wheel installations, the given path is injected to `scheme` in
`move_wheel_files()`.

For non-wheels, it is used as `--install-scripts`.

This should close #3934.