Fix pip download from an url in Windows
InstallRequirement relies on fact that setup_py_dir returns path without
directory separator at the end. In Linux this separator is stripped
by os.path.abspath, but in Windows it is not.

Also fix fix assertion message in _clean_zip_name to contain actual path
that was used in the condition.