Better document the requirements file format
Change the documentation for the requirements file format so that it
matches the implementation [0]. Also change the order of the options in
the implementation so that the documentation reads better.

Before this change the documentation included an incomplete list of
supported options.

This change adds the missing options and changes the order to match, so
that the two locations are easier to keep in sync. After this change the
list in the documentation matches SUPPORTED_OPTIONS in
src/pip/_internal/req/req_file.py

[0]: https://github.com/pypa/pip/blob/master/src/pip/_internal/req/req_file.py#L60