Catch file read errors when reading WHEEL
get_metadata delegates to the underlying implementation which tries
to locate and read the file, throwing an IOError (Python 2) or OSError
subclass on any errors.

Since the new explicit test checks the same case as brokenwheel in
test_wheel_version we remove the redundant test.

get_metadata may also throw a UnicodeDecodeError which we'll handle
next, so to preserve the existing behavior we return None on any other
exception.