Fix #7629: Restore legacy PyPy wheel interpreter tags
pip used to calculate the interpreter tag for PyPy (and other
non-CPython implementations) incorrectly. This was fixed as part
of the migration to using packaging.tags to generate the list of
potentially compatible wheel tags for a target platform.

However, there are published PyPy wheels that use the old tag
format, so this adds in a compatibility pass that accepts
additional wheel tags in cases where pip versions prior to 20.0
previously did so.