Refactor yanked warning out of package finder
The previous implementation assumes that when a best candidate is found,
it will be downloaded/installed. This will no longer be the case for the
new resolver, since it may decide to discard the candidate after
inspection and fetch a new one.

By moving it to populate_link(), we can be sure that the warning is only
set when the yanked link is actually populated for installation.