Implement Python as a dependency
If a dist contains Requires-Python metadata, it is converted into a
RequiredPythonRequirement for the resolver. This special requirement
implements find_matches() to return exactly either zero or one
candidates, depending on whether the Requires-Python metadata matches
what the user asks for.

There is one unresolved problem: LinkCandidate.get_dependencies() is
currently implemented to return an InstallRequirement collection, but a
requirement transformed from Requires-Python does not have the same
backend. How do we change the signature?