Skip test for --color=always on Windows
On Windows, colorama strips ANSI color sequences from the output and
converts them into win32 calls. So the test case does not find them in
pip's output.

We might be able to monkeypatch our way around this, but the added value
is doubtful. If the color sequences are emitted as expected on the other
platforms, then colorama should take care of the rest.