Add unit tests for colored output on Windows
These tests check if SetConsoleTextAttribute is called on Windows for
various settings of the `color` parameter to ColorizedStreamHandler.
This is achieved by monkey-patching colorama.win32. Note that this
module is an internal interface, and may change without notice.

The tests disable pytest's output capturing. Colorama checks if either
STDOUT or STDERR are a console, using GetConsoleScreenBufferInfo. With
pytest's output capturing enabled, this function fails with
ERROR_INVALID_HANDLE. A log record with an empty message is used for
testing, to avoid garbage in the test runner output.