Skip Windows test for color output unless connected to terminal
Passing color="always" won't have an effect on Windows if the output is
not a terminal. Color output on Windows relies on the win32 console
interface, and that does not apply to pipes and the like.

Colorama checks if the output is a console using
GetConsoleScreenBufferInfo with STDOUT and STDERR handles. If this check
fails, it will simply strip color sequences from the output.