Pass revision to 'hg clone' command when cloning Mercurial repository.
This ensures that only the required changesets are pulled from the server.

Some servers don’t advertise all changesets by default. An example is
Heptapod, to ensure that changesets with so-called "topics" are only cloned by
clients which know how to handle them. Passing the revision enables that these
changesets can be cloned anyway. Whether they have a "topic" or not is not
relevant to us.