Demonstrate `tmpdir` fixture issues with backslash
On some *nix systems bound to AD domains, the $USER environment is of
the form `DOMAIN\username`, which generates a tmpdir of the form:

/tmp/pytest-of-DOMAN\username/pytest-XX/test_nameX

When passed into pip commands, these paths get translated into:

/tmp/pytest-of-DOMANusername/pytest-XX/test_nameX

The backslash also breaks the `shutil.rmtree()` at the end of the
fixture, causing failures in future test runs.

This commit demonstrates the test failures that happen on such systems
by forcing `getpass.getuser()` to return a login name with a backslash.

This login name is then used by the `tmpdir*` fixtures from `py.test`
to generate per-test directory names.