Fix invalid header path when `--root` is set
When we invoke setuptools with a `--root`, the header directory we take
from the Scheme object already contains the root as a prefix. Passing
both `--root` and `--install-headers` will result in setuptools adding
the root prefix twice in the header path.

As the Scheme object already has all directories relative to `--root`,
we can pass these separately with the `--install-*` setuptools
arguments and drop the `--root` argument.

As the specific locations differ across interpreters and platforms, the
included test only checks that scripts, headers and library contents are
installed under the specified package root. This is not the case without
the fix, as some files will be installed under the duplicated root path,
outside the test `venv`.