Add support for $EDITOR with arguments
"pip config edit" doesn't work if "$EDITOR" also provides an argument for the editor. This commit fixes that.
The common example is "EDITOR=code -w", which is useful to have vscode as the editor for "git commit".
The "-w" ensures that the "code" process doesn't exit until the tab for the edited file is closed.