Add pytest-html and collect with Zuul as artifact
Add pytest-html to the test requirements so runs can provide a HTML
output.

Modify the Zuul CI configuration to produce the standalone HTML output
and update the job to collect it and mark it as an artifact.  This can
be seen by clicking on the "artificats" tab on a build result
(e.g. https://zuul.opendev.org/t/pypa/build/9f96c7aa54804fa79850b19709d1b0db).

Also add the fetch-tox-output role that collects various tox logs that
are sometimes useful.