Reorder and revise installation docs
An additional section is added to instruct the user to self-check
whether pip is installed before proceeding.

The Linux package manager section is moved to the front to encourage
users to try that first before learning about get-pip.py, which tends to
provide seemingly working but subtly broken results there. A paragraph
is added to the section to instruct users to report issues to package
managers instead.

A new section on enturepip is added as an alternative to get-pip.py,
which is almost guarenteed to be available at this point.