Pipe stderr to stdout
When working with large amounts of data, Git reports on stderr instead of
stdout. For some reason, on Git for Windows (I have not been able to reproduce
this on Linux), this can cause the subprocess to completely stall while asking
for a return from the stdout. In the context of a `pip install git+https://`,
this results in the Clone step freezing, without providing any errors or
context about what's happening (or what has gone wrong).