Add --no-backtracking option for new resolver
Adds a new option `--no-backtracking` to the new resolver. Off by
default (no change to default behaviour), enabling this option prevents
us from running the backtracking behaviour behaviour when pinning fails
and instead skip directly to the error reporing.

This can be especially useful in a CI situation, where the developer is
interested in knowing about conflicts up front in order to manually
solve them, or is worried about allowing for a potentially large amount
of time spent in backtracking.