Add separator for log file, shorten timestamps, use . for msecs
After this change, the log file looks like this.

    23:59:51.143 --- Logging started 2021-01-14 20:59:51.143 ---
    23:59:51.163 Created temporary directory: /tmp/pip-ephem-wheel-cache-jjwj2b28

Because the file log is constantly appended, the separator line
allows to quickly find the next run. Date info is moved into the
separator line to free more horizonal space for each line. The
`msec` format was modified, because the default in Python 3 is
using comma.

https://github.com/python/cpython/blob/ddc0fa3a1c1da32359f2019d14cbcfc8eb73498b/Lib/logging/__init__.py#L576