Prioritize package lookup according to their source
Previously, installation candidates would (roughly) be ordered by
version only.
With this change, a source priority is introduced, making sources behave
as fallbacks to each other: if a candidate is found in a higher priority
source, it will be downloaded from that source even if a higher version
of the package is found in another source.
Sources are ordered from highest to lowest priority as: find_links
archives, find_links links, index urls in order.